backend/orders_service/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.3
psycopg2-binary==2.9.5
requests==2.28.2
Werkzeug<3.0.0
flask-cors==4.0.0